Sparse Matrix.c
This is a code for the Sparse Matrix (A matrix in which most of the elements are Zero) in C language. It returns the Optimal matrix (Returns row, col, and corresponding value for the same). So the Zero values are removed from our final matrix. Thus Optimizing the space complexity to a great extent.